+++ engine/src/engine/search/negamax.rs
@@ -85,7 +85,7 @@ pub fn negamax(
             return if in_check {
                 Eval::DRAW
             } else {
-                eval::eval(&mut ctx.nnue, game.player)
+                eval::eval(ctx.nnue, game.player)
             };
         }
     }
@@ -155,7 +155,7 @@ pub fn negamax(
         match tt_entry {
             Some(ref e) if e.eval != Eval::NONE => e.eval,
             _ => {
-                let e = eval::eval(&mut ctx.nnue, game.player);
+                let e = eval::eval(ctx.nnue, game.player);
 
                 ctx.tt
                     .insert(game.zobrist, NodeBound::None, None, Eval::NONE, e, 0, plies);